Implement routes
- Update header navbar to route to different pages using Link.
- Create home, about-us and error pages
- Error page will show when the path provided is not valid
