+++ src/components/navbar/Header.jsx
@@ -1,19 +1,22 @@
 import React from 'react';
 import { Navbar, Nav, Container } from 'react-bootstrap';
+import { Link } from 'react-router-dom';
 
 export const Header = () => {
 	return (
-		<Navbar collapseOnSelect expand="md" bg="dark" data-bs-theme="dark" className="mb-3">
-			<Container fluid>
-				<Navbar.Brand href="#home">Meal Maker</Navbar.Brand>
-				<Navbar.Toggle aria-controls="responsive-navbar-nav" />
-				<Navbar.Collapse id="responsive-navbar-nav" className="justify-content-end">
-					<Nav>
-						<Nav.Link href="#home">Home</Nav.Link>
-						<Nav.Link href="#about-us">About Us</Nav.Link>
-					</Nav>
-				</Navbar.Collapse>
-			</Container>
-		</Navbar>
+		<header>
+			<Navbar collapseOnSelect expand="md" bg="dark" data-bs-theme="dark" className="mb-3">
+				<Container fluid>
+					<Link to="/" className="navbar-brand">Meal Maker</Link>
+					<Navbar.Toggle aria-controls="responsive-navbar-nav" />
+					<Navbar.Collapse id="responsive-navbar-nav" className="justify-content-end">
+						<Nav>
+							<Link to="/" className="nav-link">Home</Link>
+							<Link to="/about-us" className="nav-link">About Us</Link>
+						</Nav>
+					</Navbar.Collapse>
+				</Container>
+			</Navbar>
+		</header>
 	);
 };
