+++ src/components/views/AboutUsPage.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const AboutUsPage = () => {
+	return <div>AboutUsPage</div>;
+};
