+++ src/index.js
@@ -8,7 +8,7 @@ import './bootstrap.css';
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
 	<React.StrictMode>
-		<App />
+		<App/>
 	</React.StrictMode>
 );
